server/api/requirements.txt
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
asyncpg==0.31.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
dnspython==2.8.0
email-validator==2.3.0
fast-channels==1.1.0
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.9.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iso8601==2.1.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
pypika-tortoise==0.6.5
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.4.0
requests==2.33.0
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.56.0
shellingham==1.5.4
starlette==1.0.0
tortoise-orm==1.1.7
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
